added some changes

+++ src/main/java/com/example/UberBackend/service/LocationService.java
@@ -6,6 +6,6 @@ import com.example.UberBackend.dto.DriverLocationDTO;
 import java.util.List;
 
 public interface LocationService {
-    Boolean saveDriverLocation(Integer driverId ,Double latitude , Double longitude);
+    Boolean saveDriverLocation(Long driverId ,Double latitude , Double longitude);
     List<DriverLocationDTO> getNearByDrivers(Double latitude , Double longtitude , Double radius);
 }
